+++ CHANGELOG.md
@@ -1,7 +1,13 @@
+## 1.1.0
+
+- Fixed Type error bug on `MultiSelectField()` when initial options were null
+- Added `isSelected` argument on builder functions (`optionBuilder` and `childBuilder`)
+- Added `selectedOption` property to `SelectFieldMenuController()`
+
 ## 1.0.0
 
-- MultiSelectField() widget added
-- Removed customControl property from SelectFieldMenuController()
+- `MultiSelectField()` widget added
+- Removed `customControl` property from `SelectFieldMenuController()`
 
 ## 0.1.3
 
@@ -9,7 +15,7 @@
 
 ## 0.1.2
 
-- Added Material widget as a parent to OptionsMenu widget
+- Added `Material` widget as a parent to `OptionsMenu` widget
 
 ## 0.1.1